# Optimize transcriptome simulation