--- a/src/simulator.py
+++ b/src/simulator.py
@@ -383,8 +383,9 @@ def read_profile(ref_g, number_list, model_prefix, per, mode, strandness, ref_t=
                 parts = line.split("\t")
                 transcript_id = parts[0].split(".")[0]
                 tpm = float(parts[2])
-                if transcript_id.startswith("ENS") and tpm > 0:
+                if tpm > 0:
                     dict_exp[transcript_id] = tpm
+        assert len(dict_exp) > 0
         # create the ecdf dict considering the expression profiles
         ecdf_length_list, ecdf_weight_list = make_cdf(dict_exp, seq_len)