--- a/src/simulator.py
+++ b/src/simulator.py
@@ -381,11 +381,16 @@ def read_profile(ref_g, number_list, model_prefix, per, mode, strandness, ref_t=
             header = exp_file.readline()
             for line in exp_file:
                 parts = line.split("\t")
+                if len(parts) < 3:
+                    sys.stderr.write("Expression profile must contain 3 columns: ID, count, TPM \n")
+                    sys.exit(1)
                 transcript_id = parts[0].split(".")[0]
                 tpm = float(parts[2])
                 if tpm > 0:
                     dict_exp[transcript_id] = tpm
-        assert len(dict_exp) > 0
+        if len(dict_exp) == 0:
+            sys.stderr.write("Expression profile contains no TPM values > 0\n")
+            sys.exit(1)
         # create the ecdf dict considering the expression profiles
         ecdf_length_list, ecdf_weight_list = make_cdf(dict_exp, seq_len)