- consolidated code for progress updates for number of read simulated in metagenome/transcriptome/genome modes to a new method `check_print_progress`
- reduced frequency of writing updates to stdout, e.g. once every 10,000 reads simulated instead of 100
- optmized code for transcriptome simulation

Simulated 500,000 reads using 4 threads on the same machine for 3 trials:

| trial | before     | optimized | 
|-------|------------|-----------|
| 1     | 1h:00m:13s | ~~41m:42s~~ 16m:38s  |  
| 2     | 1h:12m:52s | ~~50m:47s~~ 16m:40s  | 
| 3     | 1h:03m:18s | ~~44m:53s~~ 16m:42s  | 

About ~75% reduction in **wallclock** runtime.