@kmnip 2022-04-02
I have further optimized the code in afe3cbe.

A new sample is only drawn from the KDE if the existing sample was previously used to simulate a read for the candidate transcript.

I have repeated my benchmarking trials on the same machine and updated the table above. The reduction in runtime is now ~75% (instead of ~30%)!