@SaberHQ 2022-04-06
> I have further optimized the code in [afe3cbe](https://github.com/bcgsc/NanoSim/commit/afe3cbea8d1af55cbddc0d72a6240c859d13cfb0).
> 
> A new sample is only drawn from the KDE if the existing sample was previously used to simulate a read for the candidate transcript.
> 
> I have repeated my benchmarking trials on the same machine and updated the table above. The reduction in runtime is now ~75% (instead of ~30%)!

This is really a neat and clever optimization @kmnip. Thanks for this. I also benchmarked it by simulating 1 million and 25 million read sets and I do confirm that it improved the runtime. I will add results here as well.