@SaberHQ 2022-04-08
Thank you @kmnip for the optimization and analysis. I also benchmarked it by simulating 1 million and 25 million reads using 38 threads on a same machine and the results demonstrates a reduction in wallclock runtime.

| Number of reads | before     | optimized | 
|-------|------------|-----------|
| 1 million     | 15m:56s |  10m:29s  |  
| 25 million    | 29h:46m:26s |  5h:34m:16s  |